Improve DatasetUtil and add tests
For single-fold evaluation (#41), we need to be able to slice a list of patients by a given array of randomly selected indices. Therefore, add a getRandomIndices method to get a (deterministic) list of pseudo-random integers. Also, allow slice to be operated on a longer list, but restricted by initial index and length.