Remove unused code (#20)
String comparison must be done with .equals() so that the comparison is not done at the object level.

However, if done correctly here, it would lead to a NullPointerException in getFirstToken(). getFirstToken() is actually able to correctly handle empty lines and will return a null string in that case, which is properly handled by the calling code.

Therefore, remove unused code.