Remove `NNCrossValidator` and classes that used it (refs #73)
`NNCrossValidator` is identical to `CrossValidator` except by a reset of `trained` variable on the (hardcoded) `BILSTMC3GClassifier` class. However, this is not need after 0f233b6c710f56457318f484d6779c0ad36a296b, when a new model is always trained if no cache is found. So, remove it.

Remove also `NNBestClassifierRunner`, which is then a duplicate of `NNClassifierRunner`.

Remove finally the `trained` variable and getters/setters.