Properly truncate short narratives during prediction
When training, we learn `maxLength`, which is the largest number of sentences found on a single narrative.

When predicting, we initialize a feature vector with the same size as the document to be predicted *or* `maxLength`, whatever is *smaller*. This avoids filling the feature vector with invalid data in the case of a too short document, thus leading to a wrong prediction. This also avoids initializing a feature vector which is longer than the longest training sample.

This improves n2c2 accuracy as such:
Training: 66.45% -> 96.08%
Test: 68.34% -> 74.42%

Furthermore, training accuracy now correctly fits the stop criterion of 0.5, as evaluated by the dl4j library.

This was reported upstream on deeplearning4j/dl4j-examples#761.