Save models at each epoch (fixes #87)
Previous code would save a model only after converging, which was not the original intent. Move the `saveModel(epoch)` call to inner training loop. As a consequence, save in a properties file the epoch of the best model and use it when loading the network from file.

Also, as character trigram maps do not change per epoch, save them in a new `saveParams` method called only once.

Finally, remove the `trainCounter` variable, which seemed unused.

Add tests to ensure variables and prediction are working as expected.