DRY `getLongestSequenceSize` method
Both `SentenceIterator` (`calculateMaxSentences`) and `TokenIterator` (`initializeTruncateLength`) implemented a method to calculate the longest sequence in training data and used it to initialize `truncateLength`.
Now that we have `getUnits`, DRY and pull up this method.

Note that `TokenIterator` method cleaned the text before tokenization, but such cleaning was not done at training/test time and was therefore a bug.

Not also `TokenIterator` calculated in the same method token/type coverage, which was used only for debugging before introduction of `fasttext`, that returns an approximate vector for out-of-dictionary tokens.